graph: Support --task option to show task hierarchy
The uftrace graph --task will display the task graph in the data.  The
task graph basically represents relation between tasks so that users
can see which tasks came from which parent.  This is especially useful
when trace data has many tasks.  Users may want to know which task is
important based on the relationship and time it runs.

The total time is a duration between when task started the first
function and finished the last function.  This includes time it was
sleeping.  The self time is the time it really ran on a cpu (= total
time - sleep time).  Note that computing sleep time requires
scheduling event information from the perf event.

If some task was created but never traced, it'll show the task info
but the times will be empty.

  $ uftrace graph --task
  ========== TASK GRAPH ==========
  # TOTAL-TIME   SELF-TIME : TASK
      9.002  m    1.041  m : [168086] chrome
                           :  +-[168130] chrome
                           :  |
      7.018  m    1.020  m :  +----[168131] chrome
                           :  |     |
      2.039  m    2.802  s :  |     +----[168200] chrome
                           :  |     |     +-[168201] chrome
      2.039  m   16.797 ms :  |     |     +-[168202] TaskSchedulerFo
      2.039  m  432.881 ms :  |     |     +-[168204] Chrome_ChildIOT
  ...

Signed-off-by: Namhyung Kim <namhyung@gmail.com>