arch/x86_64: Skip ENDBR64 insn for old capstone
Recent Intel CPU added CET instructions and GCC added ENDBR64 insn at
the beginning of every functions.  This makes dynamic tracing not
working since older versions of capstone engine doesn't support it.

As the instruction is just a NOP if not supported, we can just skip it
and proceed to the next instructions.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>